Fixed #15. basic auth will get values from ``get_site_info`` instead.
We will parse the target url to get the hostname and the port. However,
we will continue to honor user's decision should they give the plugin
the hostname and port explicilty. This is useful IF the plugin is incapable of
getting the right port or hostname. Some sites are known to have non-standard ports.
Currently this zap plugin uses the default base class implementation of ``get_site_info``
 so we don't recongize non-standard ports other than 80 and 443 for http and https,
respectively.