Added exclusion support. Fixed #26.
There are two issues: proxy exclusion should extend to spider and scanner but it doesn't. This is reported in code.google.com/p/zaproxy/issues/detail?id=748.

Second issue is when recurse is True, we must excluse the ``pathX`` in both spider ans canner if we want to skip ``pathX``. This is documented in code.google.com/p/zaproxy/issues/detail?id=749.